third_party/blink/web_tests/platform/mac-mac11/external/wpt/webvtt/api/VTTCue/constructor-expected.txt

This is a testharness.js-based test.
FAIL VTTCue(), initial values assert_equals: expected (string) "start" but got (undefined) undefined
PASS VTTCue(), bad start time
PASS VTTCue(), bad end time
PASS VTTCue(), text formatting
Harness: the test ran to completion.
